change api endpoints: color -> rgb, hue -> color - split main into multiple files
