+++ main.go
@@ -9,30 +9,21 @@ import (
 	"path/filepath"
 	"strconv"
 
+	"config"
+	"milight"
+
 	"github.com/evq/go-limitless"
 	"github.com/gin-gonic/gin"
 	"github.com/lucasb-eyer/go-colorful"
 )
 
-const (
-	BRIGHTNESS_RATIO  = 4
-	BRIGHTNESS_OFFSET = 2
-	BRIGHTNESS_MIN    = 1
-	BRIGHTNESS_MAX    = 100
-)
-
-type Config struct {
-	Port   int    `json:"port"`
-	Bridge string `json:"bridge"`
-}
-
 func main() {
 	ginMode := flag.String("mode", gin.ReleaseMode, "Gin Mode (debug, release, test)")
 	flag.Parse()
-	setMode(*ginMode)
+	SetMode(*ginMode)
 	router := gin.Default()
 
-	c, err := parseConfig("milight-daemon.conf")
+	c, err := ginconfig.ParseConfig("milight-daemon.conf")
 	if err != nil {
 		panic("failed to parse config file")
 	}
@@ -42,11 +33,11 @@ func main() {
 	controller := limitless.LimitlessController{
 		Host: c.Bridge,
 	}
-	groups := groups(&controller)
+	groups := milight.Groups(&controller)
 	controller.Groups = groups
 
 	router.POST("/on", func(c *gin.Context) {
-		id, err := parseGroup(c)
+		id, err := milight.ParseGroup(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
@@ -77,7 +68,7 @@ func main() {
 	})
 
 	router.POST("/off", func(c *gin.Context) {
-		id, err := parseGroup(c)
+		id, err := milight.ParseGroup(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
@@ -107,25 +98,25 @@ func main() {
 		c.JSON(200, msg)
 	})
 
-	router.POST("/color", func(c *gin.Context) {
-		id, err := parseGroup(c)
+	router.POST("/rgb", func(c *gin.Context) {
+		id, err := milight.ParseGroup(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
 		}
-		color, err := parseColorRGB(c)
+		rgb, err := milight.ParseRGB(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
 		}
 		msg := gin.H{
-			"command": "color",
-			"color":   color,
+			"command": "rgb",
+			"rgb":     rgb,
 			"group":   0,
 		}
 		if id == -1 {
 			for _, g := range controller.Groups {
-				err := g.SendColor(color)
+				err := g.SendColor(rgb)
 				if err != nil {
 					err = errors.New("failed to send color")
 					c.AbortWithError(500, err)
@@ -133,7 +124,7 @@ func main() {
 				}
 			}
 		} else {
-			err = controller.Groups[id].SendColor(color)
+			err = controller.Groups[id].SendColor(rgb)
 			if err != nil {
 				err = errors.New("failed to send color")
 				c.AbortWithError(500, err)
@@ -146,12 +137,12 @@ func main() {
 
 	router.POST("/brightness", func(c *gin.Context) {
 		var bl uint8
-		id, err := parseGroup(c)
+		id, err := milight.ParseGroup(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
 		}
-		bl, err = parseBrightnessLevel(c)
+		bl, err = milight.ParseBrightnessLevel(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
@@ -182,19 +173,19 @@ func main() {
 		c.JSON(200, msg)
 	})
 
-	router.POST("/hue", func(c *gin.Context) {
-		id, err := parseGroup(c)
+	router.POST("/color", func(c *gin.Context) {
+		id, err := milight.ParseGroup(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
 		}
-		color, err := parseColorName(c)
+		color, err := milight.ParseColor(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
 		}
 		msg := gin.H{
-			"command": "hue",
+			"command": "color",
 			"color":   fmt.Sprintf("%x", color),
 			"group":   0,
 		}
@@ -202,7 +193,7 @@ func main() {
 			for _, g := range controller.Groups {
 				err = g.SetHue(color)
 				if err != nil {
-					err = errors.New("failed to set hue")
+					err = errors.New("failed to set color")
 					c.AbortWithError(500, err)
 					return
 				}
@@ -210,7 +201,7 @@ func main() {
 		} else {
 			err = controller.Groups[id].SetHue(color)
 			if err != nil {
-				err = errors.New("failed to set hue")
+				err = errors.New("failed to set color")
 				c.AbortWithError(500, err)
 				return
 			}
@@ -220,7 +211,7 @@ func main() {
 	})
 
 	router.POST("/white", func(c *gin.Context) {
-		id, err := parseGroup(c)
+		id, err := milight.ParseGroup(c)
 		if err != nil {
 			c.AbortWithError(400, err)
 			return
@@ -253,121 +244,7 @@ func main() {
 	router.Run(host)
 }
 
-func parseConfig(configPath string) (Config, error) {
-	c := Config{}
-	absConfigPath, err := filepath.Abs(configPath)
-	if err != nil {
-		return c, err
-	}
-	file, _ := os.Open(absConfigPath)
-	decoder := json.NewDecoder(file)
-	err = decoder.Decode(&c)
-	if err != nil {
-		return c, err
-	}
-	return c, nil
-}
-
-func groups(c *limitless.LimitlessController) []limitless.LimitlessGroup {
-	g := make([]limitless.LimitlessGroup, 4, 4)
-	for i := 0; i < 4; i++ {
-		g[i] = limitless.LimitlessGroup{
-			Id:         i + 1,
-			Controller: c,
-		}
-	}
-	return g
-}
-
-func parseGroup(c *gin.Context) (int, error) {
-	group := c.Query("group")
-	id, err := strconv.Atoi(group)
-	if err != nil {
-		err = errors.New("failed to parse group")
-		return -1, err
-	}
-	if id < 0 || id > 4 {
-		err = errors.New("invalid id. must be <= 0 or >= 4")
-		return -1, err
-	}
-	// use id as index for groups
-	id = id - 1
-	return id, nil
-}
-
-func parseColorRGB(c *gin.Context) (colorful.Color, error) {
-	rgb := map[string]float64{
-		"r": 0,
-		"g": 0,
-		"b": 0,
-	}
-	var err error
-	for k, v := range rgb {
-		if v, err = strconv.ParseFloat(c.Query(k), 64); err != nil {
-			err = errors.New("failed to parse color")
-			return colorful.Color{}, err
-		}
-		if v < 0 || v > 255 {
-			err = errors.New("invalid color value. must be <= 0 or >= 255")
-			return colorful.Color{}, err
-		}
-		rgb[k] = v
-	}
-	fmt.Println(rgb)
-	color := colorful.Color{
-		rgb["r"] / 255.0,
-		rgb["g"] / 255.0,
-		rgb["b"] / 255.0,
-	}
-	return color, nil
-}
-
-func parseBrightnessLevel(c *gin.Context) (uint8, error) {
-	level := c.Query("level")
-	b64, err := strconv.ParseUint(level, 10, 8)
-	if err != nil {
-		err = errors.New("failed to parse brightness level")
-		return 0, err
-	}
-	b := uint8(b64)
-	if b < BRIGHTNESS_MIN || b > BRIGHTNESS_MAX {
-		err = errors.New("invalid brightness level. Must be between 1-100")
-		return 0, err
-	}
-	b = b/BRIGHTNESS_RATIO + BRIGHTNESS_OFFSET
-	return b, nil
-}
-
-func parseColorName(c *gin.Context) (uint8, error) {
-	Colors := map[string]uint8{
-		"violet":        0x00,
-		"blue":          0x10,
-		"baby_blue":     0x20,
-		"aqua":          0x30,
-		"mint":          0x40,
-		"seafoam_green": 0x50,
-		"green":         0x60,
-		"lime_green":    0x70,
-		"yellow":        0x80,
-		"yellow_orange": 0x90,
-		"orange":        0xA0,
-		"red":           0xB0,
-		"pink":          0xC0,
-		"fusia":         0xD0,
-		"lilac":         0xE0,
-		"lavendar":      0xF0,
-	}
-
-	color := c.Query("color")
-	colorHex, ok := Colors[color]
-	if !ok {
-		err := errors.New("invalid color name")
-		return 0, err
-	}
-	return colorHex, nil
-}
-
-func setMode(mode string) {
+func SetMode(mode string) {
 	switch mode {
 	case "release":
 		gin.SetMode(gin.ReleaseMode)
