+++ service/config.go
@@ -1,0 +1,21 @@
+package config
+
+type Config struct {
+	Port   int    `json:"port"`
+	Bridge string `json:"bridge"`
+}
+
+func ParseConfig(configPath string) (Config, error) {
+	c := Config{}
+	absConfigPath, err := filepath.Abs(configPath)
+	if err != nil {
+		return c, err
+	}
+	file, _ := os.Open(absConfigPath)
+	decoder := json.NewDecoder(file)
+	err = decoder.Decode(&c)
+	if err != nil {
+		return c, err
+	}
+	return c, nil
+}
